+++ adapters/Facebook/FacebookAdapter/GADMAdapterFacebook.h
@@ -14,9 +14,10 @@
 
 @import GoogleMobileAds;
 
+extern NSString *const kGADMAdapterFacebookVersion;
+
 /// Adapter for communicating with Facebook's Audience Network to fetch banner, interstitial,
 /// reward-based video, and native ads.
-@interface GADMAdapterFacebook
-    : NSObject<GADMAdNetworkAdapter, GADMRewardBasedVideoAdNetworkAdapter>
+@interface GADMAdapterFacebook : NSObject <GADMAdNetworkAdapter>
 
 @end