+++ adapters/Facebook/FacebookAdapter/GADMediationAdapterFacebook.h
@@ -1,4 +1,4 @@
-// Copyright 2016 Google Inc.
+// Copyright 2018 Google Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,15 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#import "GADFBNetworkExtras.h"
+#import <Foundation/Foundation.h>
+@import GoogleMobileAds;
 
-@implementation GADFBNetworkExtras
+extern NSString *const kGADMediationAdapterFacebookPublisherID;
 
-- (instancetype)init {
-  self = [super init];
-  if (self) {
-  }
-  return self;
-}
+@interface GADMediationAdapterFacebook : NSObject <GADMediationAdapter>
 
 @end